+++ SplunkAppForWazuh/appserver/static/js/controllers/settings/api/settingsApiCtrl.js
@@ -27,6 +27,7 @@ define(['../../module'], function (controllers) {
       this.currentDataService = $currentDataService
       this.notification = $notificationService
       this.savingApi = false
+      this.scope.runAs = false;
     }
 
     /**
@@ -38,12 +39,16 @@ define(['../../module'], function (controllers) {
       this.scope.checkManager = entry => this.checkManager(entry)
       this.scope.editEntry = entry => this.editEntry(entry)
       this.scope.removeManager = entry => this.removeManager(entry)
-      this.scope.updateEntry = (user, pass, url, port) =>
-        this.updateEntry(user, pass, url, port)
+      this.scope.updateEntry = (user, pass, url, port, runAs) =>
+        this.updateEntry(user, pass, url, port, runAs)
       this.scope.selectManager = mg => this.selectManager(mg)
-      this.scope.submitApiForm = (user, api, url, port) =>
-        this.submitApiForm(user, api, url, port)
+      this.scope.submitApiForm = (user, api, url, port, runAs) =>
+        this.submitApiForm(user, api, url, port, runAs)
+      this.scope.getIconAndTooltip = entry => this.getIconAndTooltip(entry)
       this.init()
+
+      // Init form values
+      this.clearForm();
 
       // Listens for changes in the selected API
       this.scope.$on('APIChanged', (event, key) => {
@@ -89,7 +94,7 @@ define(['../../module'], function (controllers) {
           }
         }
       } catch (err) {
-        this.notification.showErrorToast('Error loading data')
+        this.notification.showErrorToast('Error loading saved APIs')
       }
     }
 
@@ -152,6 +157,7 @@ define(['../../module'], function (controllers) {
      * Set form visible
      */
     addNewApiClick() {
+      this.clearForm()
       this.scope.showForm = !this.scope.showForm
       this.scope.edit = false
       this.scope.currentEntryKey = false
@@ -170,6 +176,7 @@ define(['../../module'], function (controllers) {
         this.scope.url = entry.url
         this.scope.pass = entry.pass
         this.scope.port = entry.portapi
+        this.scope.runAs = entry.runAs;
         this.scope.user = entry.userapi
         this.scope.managerName = entry.managerName
         this.scope.filterType = entry.filterType
@@ -184,7 +191,7 @@ define(['../../module'], function (controllers) {
      * Edits an entry
      * @param {Object} entry
      */
-    async updateEntry(user, pass, url, port) {
+    async updateEntry(user, pass, url, port, runAs = false) {
       try {
         this.scope.loadingVizz = true
         if (this.savingApi) {
@@ -198,6 +205,7 @@ define(['../../module'], function (controllers) {
         this.scope.entry.portapi = port
         this.scope.entry.userapi = user
         this.scope.entry.passapi = pass
+        this.scope.entry.runAs = runAs
         this.scope.entry.filterType = this.scope.filterType
         this.scope.entry.filterName = this.scope.filterName
         this.scope.entry.managerName = this.scope.managerName
@@ -226,7 +234,7 @@ define(['../../module'], function (controllers) {
 
         this.scope.edit = false
         this.scope.loadingVizz = false
-        this.notification.showSuccessToast('Updated API')
+        this.notification.showSuccessToast('API updated')
       } catch (err) {
         this.scope.loadingVizz = false
         this.notification.showErrorToast(err || 'Cannot update API')
@@ -259,7 +267,7 @@ define(['../../module'], function (controllers) {
     /**
      * Adds a new API
      */
-    async submitApiForm(user, pass, url, port) {
+    async submitApiForm(user, pass, url, port, runAs = false) {
       try {
         this.scope.loadingVizz = true
         this.scope.validatingError = []
@@ -291,7 +299,8 @@ define(['../../module'], function (controllers) {
           url: form_url,
           portapi: form_apiport,
           userapi: form_apiuser,
-          passapi: form_apipass
+          passapi: form_apipass,
+          runAs: runAs
         }
 
         // If connected to the API then continue
@@ -397,6 +406,7 @@ define(['../../module'], function (controllers) {
       this.scope.port = ''
       this.scope.user = ''
       this.scope.pass = ''
+      this.scope.runAs = false;
     }
 
     setYellowStar(key) {
@@ -404,6 +414,29 @@ define(['../../module'], function (controllers) {
         api['_key'] === key ? (api.selected = true) : (api.selected = false)
       })
     }
+
+
+    /**
+     * Evaluates the class and tooltip text for the given API 
+     * object depending on the `run_as` value.
+     * 
+     * If the `run_as` field is true, a check icon is shown. If it is false,
+     * a cross icon is shown instead. Respectively, the tooltip text takes 
+     * 'enabled' or 'disable'.
+     * 
+     * @param {Object} entry API object to evaluate.
+     * @returns Object with the CSS class for the icon and string text 
+     * for the tooltip.
+     */
+    getIconAndTooltip(entry) {
+      // Cast to boolean
+      const runAs = (entry.runAs === "true") || (entry.runAs === true)
+      return {
+        class: runAs ? 'fa fa-check' : 'fa fa-times',
+        tooltip: runAs ? 'enabled' : 'disabled'
+      }
+    }
+
   }
   controllers.controller('settingsApiCtrl', SettingsApi)
 })